Made ASM work (and also better)
